pyproject.toml
[project]
name = "microshop"
version = "0.1.0"
description = ""
authors = [
    {name = "Norix",email = "121669280+Norix10@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn[standart] (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]

[tool.poetry]
packages = [
    { include = "microshop" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"